fix(games): store crashPoint/cashoutMultiplier as double precision
Postgres 'real' is 32-bit, JS numbers are 64-bit doubles - saving and
reloading a round was silently losing precision, which would make
ProvablyFairResult.verify() fail on a perfectly legitimate round.
Confirmed with a save/reload round-trip before and after the fix.

+++ services/games/src/infrastructure/persistence/entities/bet.entity.ts
@@ -19,6 +19,6 @@ export class BetEntity {
   @Enum(() => BetStatus)
   status!: BetStatus;
 
-  @Property({ type: 'float', nullable: true })
+  @Property({ type: 'double', nullable: true })
   cashoutMultiplier?: number | null;
 }
